services/api/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1          # passlib 1.7.4 is incompatible with bcrypt>=5.0
python-multipart==0.0.20

# Config & validation
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# File I/O
aiofiles==24.1.0

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
